Create 541(Sale_Season)
Sale Season

It's the sale season again and Chef bought items worth a total of X rupees. The sale season offer is as follows:

If X≤100, no discount.
If 100<X≤1000, discount is 25 rupees.
If 1000<X≤5000, discount is 100 rupees.
If X>5000, discount is 500 rupees.

Find the final amount Chef needs to pay for his shopping.