Create 793(Degree_of_Polynomial)
Degree of Polynomial

In mathematics, the degree of polynomials in one variable is the highest power of the variable in the algebraic expression with non-zero coefficient.

Chef has a polynomial in one variable x with N terms. Find the degree of the polynomial.

Note: It is guaranteed that there exists at least one term with non-zero coefficient.